[10/02/2012] M    SAT/Base.h 1. [MOD] Change the type of eName in _SetLong from EPAramter to int.
M    SAT/IsotropicWaveletSAT.h
1. [MOD] Change the type of eName in _SetLong from EPAramter to int.
2. [ADD] Add this-> to the members of CBase<T>.
3. [ADD] Add CBase<T> to the method of CBase<T>.

M    SAT/SimpleND.h
1. [MOD] Use Wavelet::CBase as the inherited class.

M    SAT/WaveletSAT.h
1. [MOD] Change the type of eName in _SetLong from EPAramter to int.
 
M    SimpleNDWithNrrd/SimpleNDWithNrrd.cpp
1. [DEL] Remove ^M.

+++ SimpleNDWithNrrd/SimpleNDWithNrrd.cpp
@@ -293,7 +293,7 @@ main(int argn, char* argv[])
 				#else		// MOD-BY-LEETEN 10/01/2012-TO:
 				for(size_t b = 0; b < uNrOfBins; b++)
 					if( fabs(vdH[b]) > dThreshold )
-						printf( "\t%d:%+.2f\n", b, vdH[b]);
+					  printf( "\t%d:%+.2f\n", (unsigned int)b, vdH[b]); // MOD-BY-LEETEN 10/02/2012-FROM: printf( "\t%d:%+.2f\n", b, vdH[b]);
 				#endif		// MOD-BY-LEETEN 10/01/2012-END
 				printf("E:%f\n", dError);
 			}
